+++ projectroles/tests/test_remote_projects_api.py
@@ -1,4 +1,5 @@
 """Test for the remote projects API in the projectroles app"""
+from datetime import datetime as dt
 import uuid
 
 from django.conf import settings
@@ -310,17 +311,8 @@ class TestSyncSourceData(
 
         self.remote_api = RemoteProjectAPI()
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
-    def test_sync_create(self):
-        """Test sync with non-existing project data and READ_ROLE access"""
-
-        # Assert preconditions
-        self.assertEqual(Project.objects.all().count(), 0)
-        self.assertEqual(RoleAssignment.objects.all().count(), 0)
-        self.assertEqual(User.objects.all().count(), 1)
-        self.assertEqual(RemoteProject.objects.all().count(), 0)
-
-        remote_data = {
+        # Default data to receive from source when testing site in target mode
+        self.default_data = {
             'users': {
                 SOURCE_USER_UUID: {
                     'sodar_uuid': SOURCE_USER_UUID,
@@ -364,6 +356,17 @@ class TestSyncSourceData(
             }
         }
 
+    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
+    def test_sync_create(self):
+        """Test sync with non-existing project data and READ_ROLE access"""
+
+        # Assert preconditions
+        self.assertEqual(Project.objects.all().count(), 0)
+        self.assertEqual(RoleAssignment.objects.all().count(), 0)
+        self.assertEqual(User.objects.all().count(), 1)
+        self.assertEqual(RemoteProject.objects.all().count(), 0)
+
+        remote_data = self.default_data
         update_data = self.remote_api.sync_source_data(
             self.source_site, remote_data)
 
@@ -421,6 +424,30 @@ class TestSyncSourceData(
             'sodar_uuid': uuid.UUID(SOURCE_PROJECT_ROLE_UUID)}
         self.assertEqual(model_to_dict(p_owner_obj), expected)
 
+        remote_cat_obj = RemoteProject.objects.get(
+            site=self.source_site,
+            project_uuid=category_obj.sodar_uuid)
+        expected = {
+            'id': remote_cat_obj.pk,
+            'site': self.source_site.pk,
+            'project_uuid': category_obj.sodar_uuid,
+            'level': REMOTE_LEVEL_READ_ROLES,
+            'date_access': remote_cat_obj.date_access,
+            'sodar_uuid': remote_cat_obj.sodar_uuid}
+        self.assertEqual(model_to_dict(remote_cat_obj), expected)
+
+        remote_project_obj = RemoteProject.objects.get(
+            site=self.source_site,
+            project_uuid=project_obj.sodar_uuid)
+        expected = {
+            'id': remote_project_obj.pk,
+            'site': self.source_site.pk,
+            'project_uuid': project_obj.sodar_uuid,
+            'level': REMOTE_LEVEL_READ_ROLES,
+            'date_access': remote_project_obj.date_access,
+            'sodar_uuid': remote_project_obj.sodar_uuid}
+        self.assertEqual(model_to_dict(remote_project_obj), expected)
+
         # Assert update_data changes
         expected = dict(remote_data)
         expected['projects'][SOURCE_CATEGORY_UUID]['status'] = 'created'
@@ -429,4 +456,172 @@ class TestSyncSourceData(
         expected['projects'][SOURCE_PROJECT_UUID]['status'] = 'created'
         expected['projects'][SOURCE_PROJECT_UUID]['roles'][
             SOURCE_PROJECT_ROLE_UUID]['status'] = 'created'
+        self.assertEqual(update_data, expected)
+
+    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
+    def test_sync_update(self):
+        """Test sync with existing project data and READ_ROLE access"""
+
+        # Set up target category and project
+        category_obj = self._make_project(
+            title='NewCategoryTitle',
+            type=PROJECT_TYPE_CATEGORY,
+            parent=None,
+            description='New description',
+            readme='New readme',
+            sodar_uuid=SOURCE_CATEGORY_UUID)
+        project_obj = self._make_project(
+            title='NewProjectTitle',
+            type=PROJECT_TYPE_PROJECT,
+            parent=category_obj,
+            description='New description',
+            readme='New readme',
+            sodar_uuid=SOURCE_PROJECT_UUID)
+
+        # Set up user and roles
+        target_user = self._make_sodar_user(
+            username=SOURCE_USER_USERNAME,
+            name='NewFirstName NewLastName',
+            first_name='NewFirstName',
+            last_name='NewLastName')
+        c_owner_obj = self._make_assignment(
+            category_obj, target_user, self.role_owner)
+        p_owner_obj = self._make_assignment(
+            project_obj, target_user, self.role_owner)
+
+        # Set up RemoteProject objects
+        remote_cat = self._make_remote_project(
+            project_uuid=category_obj.sodar_uuid,
+            site=self.source_site,
+            level=REMOTE_LEVEL_READ_ROLES)
+        remote_project = self._make_remote_project(
+            project_uuid=project_obj.sodar_uuid,
+            site=self.source_site,
+            level=REMOTE_LEVEL_READ_ROLES)
+
+        # Assert preconditions
+        self.assertEqual(Project.objects.all().count(), 2)
+        self.assertEqual(RoleAssignment.objects.all().count(), 2)
+        self.assertEqual(User.objects.all().count(), 2)
+        self.assertEqual(RemoteProject.objects.all().count(), 2)
+
+        remote_data = self.default_data
+
+        # Add new user and contributor role to source project
+        new_user_username = 'newuser@' + SOURCE_USER_DOMAIN
+        new_user_uuid = str(uuid.uuid4())
+        new_role_uuid = str(uuid.uuid4())
+
+        remote_data['users'][str(new_user_uuid)] = {
+            'sodar_uuid': new_user_uuid,
+            'username': new_user_username,
+            'name': 'Some Name',
+            'first_name': 'Some',
+            'last_name': 'Name',
+            'email': 'newuser@example.com',
+            'groups': [SOURCE_USER_GROUP]}
+        remote_data['projects'][SOURCE_PROJECT_UUID][
+            'roles'][new_role_uuid] = {
+                'user': new_user_username,
+                'role': PROJECT_ROLE_CONTRIBUTOR}
+
+        update_data = self.remote_api.sync_source_data(
+            self.source_site, remote_data)
+
+        # Assert database status
+        self.assertEqual(Project.objects.all().count(), 2)
+        self.assertEqual(RoleAssignment.objects.all().count(), 3)
+        self.assertEqual(User.objects.all().count(), 3)
+        self.assertEqual(RemoteProject.objects.all().count(), 2)
+
+        target_user = User.objects.get(username=SOURCE_USER_USERNAME)
+        new_user = User.objects.get(username=new_user_username)
+
+        category_obj.refresh_from_db()
+        expected = {
+            'id': category_obj.pk,
+            'title': SOURCE_CATEGORY_TITLE,
+            'type': PROJECT_TYPE_CATEGORY,
+            'description': SOURCE_PROJECT_DESCRIPTION,
+            'parent': None,
+            'submit_status': SUBMIT_STATUS_OK,
+            'sodar_uuid': uuid.UUID(SOURCE_CATEGORY_UUID)}
+        model_dict = model_to_dict(category_obj)
+        model_dict.pop('readme', None)
+        self.assertEqual(model_dict, expected)
+
+        c_owner_obj.refresh_from_db()
+        expected = {
+            'id': c_owner_obj.pk,
+            'project': category_obj.pk,
+            'user': target_user.pk,
+            'role': self.role_owner.pk,
+            'sodar_uuid': c_owner_obj.sodar_uuid}
+        self.assertEqual(model_to_dict(c_owner_obj), expected)
+
+        project_obj.refresh_from_db()
+        expected = {
+            'id': project_obj.pk,
+            'title': SOURCE_PROJECT_TITLE,
+            'type': PROJECT_TYPE_PROJECT,
+            'description': SOURCE_PROJECT_DESCRIPTION,
+            'parent': category_obj.pk,
+            'submit_status': SUBMIT_STATUS_OK,
+            'sodar_uuid': uuid.UUID(SOURCE_PROJECT_UUID)}
+        model_dict = model_to_dict(project_obj)
+        model_dict.pop('readme', None)
+        self.assertEqual(model_dict, expected)
+
+        p_owner_obj.refresh_from_db()
+        expected = {
+            'id': p_owner_obj.pk,
+            'project': project_obj.pk,
+            'user': target_user.pk,
+            'role': self.role_owner.pk,
+            'sodar_uuid': p_owner_obj.sodar_uuid}
+        self.assertEqual(model_to_dict(p_owner_obj), expected)
+
+        p_contrib_obj = RoleAssignment.objects.get(
+            project__sodar_uuid=SOURCE_PROJECT_UUID,
+            role__name=PROJECT_ROLE_CONTRIBUTOR)
+        expected = {
+            'id': p_contrib_obj.pk,
+            'project': project_obj.pk,
+            'user': new_user.pk,
+            'role': self.role_contributor.pk,
+            'sodar_uuid': p_contrib_obj.sodar_uuid}
+        self.assertEqual(model_to_dict(p_contrib_obj), expected)
+
+        remote_cat_obj = RemoteProject.objects.get(
+            site=self.source_site,
+            project_uuid=category_obj.sodar_uuid)
+        expected = {
+            'id': remote_cat_obj.pk,
+            'site': self.source_site.pk,
+            'project_uuid': category_obj.sodar_uuid,
+            'level': REMOTE_LEVEL_READ_ROLES,
+            'date_access': remote_cat_obj.date_access,
+            'sodar_uuid': remote_cat_obj.sodar_uuid}
+        self.assertEqual(model_to_dict(remote_cat_obj), expected)
+
+        remote_project_obj = RemoteProject.objects.get(
+            site=self.source_site,
+            project_uuid=project_obj.sodar_uuid)
+        expected = {
+            'id': remote_project_obj.pk,
+            'site': self.source_site.pk,
+            'project_uuid': project_obj.sodar_uuid,
+            'level': REMOTE_LEVEL_READ_ROLES,
+            'date_access': remote_project_obj.date_access,
+            'sodar_uuid': remote_project_obj.sodar_uuid}
+        self.assertEqual(model_to_dict(remote_project_obj), expected)
+
+        # Assert update_data changes
+        expected = dict(remote_data)
+        expected['users'][SOURCE_USER_UUID]['status'] = 'updated'
+        expected['users'][new_user_uuid]['status'] = 'created'
+        expected['projects'][SOURCE_CATEGORY_UUID]['status'] = 'updated'
+        expected['projects'][SOURCE_PROJECT_UUID]['status'] = 'updated'
+        expected['projects'][SOURCE_PROJECT_UUID]['roles'][
+            new_role_uuid]['status'] = 'created'
         self.assertEqual(update_data, expected)